Add ExceptionHandler. Problem solution required: must return the object of the Сandidate registered in the database correctly
 Package com.exadel.backendservice.controllers.CandidateController -> Method ResponseEntity<String>	registerCandidate (RegisterCandidateDto candidateDto)
