+++ src/main/java/com/exadel/backendservice/entity/Country.java
@@ -1,7 +1,9 @@
 package com.exadel.backendservice.entity;
 
-import lombok.*;
-import lombok.experimental.FieldDefaults;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
