--- src/main/java/com/exadel/backendservice/exception/ApiRequestException.java
@@ -1,11 +1,0 @@
-package com.exadel.backendservice.exception;
-
-public class ApiRequestException extends RuntimeException{
-    public ApiRequestException(String message) {
-        super(message);
-    }
-
-    public ApiRequestException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
