+++ src/main/java/com/exadel/backendservice/exception/ApiRequestExceptionDto.java
@@ -1,0 +1,11 @@
+package com.exadel.backendservice.exception;
+
+public class ApiRequestExceptionDto extends RuntimeException{
+    public ApiRequestExceptionDto(String message) {
+        super(message);
+    }
+
+    public ApiRequestExceptionDto(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
