--- src/main/java/com/exadel/backendservice/exception/ApiResponseException.java
@@ -1,11 +1,0 @@
-package com.exadel.backendservice.exception;
-
-public class ApiResponseException extends RuntimeException {
-    public ApiResponseException(String message) {
-        super(message);
-    }
-
-    public ApiResponseException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
