+++ src/main/java/com/exadel/backendservice/exception/ApiResponseExceptionDto.java
@@ -1,0 +1,11 @@
+package com.exadel.backendservice.exception;
+
+public class ApiResponseExceptionDto extends RuntimeException {
+    public ApiResponseExceptionDto(String message) {
+        super(message);
+    }
+
+    public ApiResponseExceptionDto(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
